Teste de inserção de Partido a partir de CSV funcionando.

+++ dados-parlamentares/src/main/java/br/uel/bd1/dadosparlamentares/business/BlocoBusiness.java
@@ -3,6 +3,7 @@ package br.uel.bd1.dadosparlamentares.business;
 import br.uel.bd1.dadosparlamentares.dao.GenericDAO;
 import br.uel.bd1.dadosparlamentares.model.Bloco;
 import org.supercsv.cellprocessor.ParseInt;
+import org.supercsv.cellprocessor.ParseLong;
 import org.supercsv.cellprocessor.constraint.NotNull;
 import org.supercsv.cellprocessor.ift.CellProcessor;
 
@@ -13,9 +14,9 @@ public class BlocoBusiness extends GenericBusiness<Bloco> {
     @Override
     protected CellProcessor[] getProcessors() {
         final CellProcessor[] processors = new CellProcessor[] {
-            new NotNull(new ParseInt()),
+            new NotNull(new ParseLong()),
             new NotNull(),
-            new NotNull(new ParseInt())
+            new NotNull(new ParseLong())
         };
 
         return processors;
